requirement.txt
asgiref==3.3.4
Django==3.2
django-next-prev==1.1.0
pytz==2021.1
sqlparse==0.4.1
